requirements.txt
matplotlib==3.7.0
pandas==1.5.3
pyserial==3.5